I1-55 feat: add decorators: wrap story in div with font
